+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## 0.1.0
+
+- `onEmit`, `onRequest`, `emitEvent`, `sendRequest`